Issue #191 - Split /api/health into a liveness probe and a seed-data check
/api/health answered two unrelated questions in one request — is the app
alive, and does the DB still hold the reference rows the code depends on —
and the Compose healthcheck could not tell them apart. A fresh Finances,
whose transaction_categories are user-created, therefore reported 503 and
marked the container unhealthy while serving perfectly (#178).

/api/health is now liveness only: one SELECT 1 plus the build stamp. The
drift check moves to /api/health/seed-data with its payloads preserved
verbatim, gated by auth() in the handler with a 401 — not via proxy.ts's
matcher, which answers a denied request with a 302 to /login, so a monitor
following redirects would read a 200 HTML page as success.

The release smoke test drops the FINANCE_APP_DB: Finances_Test workaround
carried since v0.1.4 and boots a fresh empty Finances, plus a new assertion
that Docker reports the container healthy.

Also fixes a defect found while verifying this, which blocked two of the
acceptance criteria: the healthcheck probed http://localhost:3001, /etc/hosts
maps localhost to both 127.0.0.1 and ::1, BusyBox wget tries IPv6 first, and
the Next standalone server binds IPv4 only — so the probe never connected and
finance-app was permanently unhealthy whatever the route answered.
