+++ app/app/api/health/route.ts
@@ -1,53 +1,14 @@
 import { db } from "@/lib/db";
 import { sql } from "drizzle-orm";
-import { SEED_REFERENCES } from "@/lib/constants/reference-ids";
 import { BUILD_INFO } from "@/lib/version";
-import { reportError } from "@/lib/report";
 
-export interface DriftEntry {
-  table: string;
-  id: number;
-  expected: string;
-  actual: string | null;
-}
-
-export async function checkSeedReferences(): Promise<DriftEntry[]> {
-  const drift: DriftEntry[] = [];
-
-  for (const group of SEED_REFERENCES) {
-    const ids = group.expected.map((r) => r.id);
-
-    const result = await db.execute(sql`
-      SELECT ${sql.identifier(group.idColumn)} AS id,
-             ${sql.identifier(group.nameColumn)} AS name
-      FROM ${sql.identifier(group.table)}
-      WHERE ${sql.identifier(group.idColumn)} IN (${sql.join(
-        ids.map((id) => sql`${id}`),
-        sql`, `,
-      )})
-    `);
-
-    const actualById = new Map<number, string>();
-    for (const row of result.rows as { id: number; name: string }[]) {
-      actualById.set(row.id, row.name);
-    }
-
-    for (const expected of group.expected) {
-      const actual = actualById.get(expected.id) ?? null;
-      if (actual !== expected.name) {
-        drift.push({
-          table: group.table,
-          id: expected.id,
-          expected: expected.name,
-          actual,
-        });
-      }
-    }
-  }
-
-  return drift;
-}
-
+// Liveness only (#191): one query, no seed-row inspection. This is the target of
+// the Compose healthcheck and the release smoke test, and it is one of the few
+// public routes, so it stays cheap enough to poll every 10s and discloses nothing
+// beyond the build stamp. Seed-row drift moved to /api/health/seed-data — it is a
+// data problem a container restart cannot fix, so it must not drive restart policy,
+// and a fresh Finances (no user-created transaction_categories) legitimately trips
+// it while being perfectly alive.
 export async function GET() {
   try {
     await db.execute(sql`SELECT 1`);
@@ -58,28 +19,9 @@ export async function GET() {
     );
   }
 
-  let drift: DriftEntry[];
-  try {
-    drift = await checkSeedReferences();
-  } catch (err) {
-    reportError(err, { route: "/api/health" });
-    return Response.json(
-      { status: "error", db: "connected", seedData: "error" },
-      { status: 503 },
-    );
-  }
-
-  if (drift.length > 0) {
-    return Response.json(
-      { status: "error", db: "connected", seedData: "drift", drift },
-      { status: 503 },
-    );
-  }
-
   return Response.json({
     status: "ok",
     db: "connected",
-    seedData: "ok",
     build: BUILD_INFO,
   });
 }
